+++ ios/events/TrueSheetLifecycleEvents.mm
@@ -66,6 +66,14 @@
   emitter->onDidDismiss({});
 }
 
++ (void)emitDismissAttempt:(std::shared_ptr<const facebook::react::EventEmitter>)eventEmitter {
+  if (!eventEmitter)
+    return;
+
+  auto emitter = std::static_pointer_cast<TrueSheetViewEventEmitter const>(eventEmitter);
+  emitter->onDismissAttempt({});
+}
+
 @end
 
 #endif  // RCT_NEW_ARCH_ENABLED